Refactoring in preparation for AMP support
