+++ docs/tools/build.py
@@ -9,7 +9,6 @@ import subprocess
 import sys
 import time
 
-import bs4
 import jinja2
 import livereload
 import markdown.util
@@ -18,10 +17,11 @@ from mkdocs import config
 from mkdocs import exceptions
 from mkdocs.commands import build as mkdocs_build
 
-from concatenate import concatenate
-
+import amp
 import mdx_clickhouse
 import nav
+import redirects
+import single_page
 import test
 import util
 import website
@@ -144,18 +144,19 @@ def build_for_lang(lang, args):
             edit_uri=f'edit/master/docs/{lang}',
             markdown_extensions=markdown_extensions,
             plugins=plugins,
-            extra={
-                'stable_releases': args.stable_releases,
-                'version_prefix': args.version_prefix,
-                'single_page': False,
-                'rev': args.rev,
-                'rev_short': args.rev_short,
-                'rev_url': args.rev_url,
-                'website_url': website_url,
-                'events': args.events,
-                'languages': languages,
-                'includes_dir':  os.path.join(os.path.dirname(__file__), '..', '_includes')
-            }
+            extra=dict(
+                stable_releases=args.stable_releases,
+                version_prefix=args.version_prefix,
+                single_page=False,
+                rev=args.rev,
+                rev_short=args.rev_short,
+                rev_url=args.rev_url,
+                website_url=website_url,
+                events=args.events,
+                languages=languages,
+                includes_dir=os.path.join(os.path.dirname(__file__), '..', '_includes'),
+                is_amp=False
+            )
         )
 
         if os.path.exists(config_path):
@@ -174,8 +175,11 @@ def build_for_lang(lang, args):
                 mdx_clickhouse.PatchedMacrosPlugin.disabled = True
                 mkdocs_build.build(cfg)
 
+        if not (args.skip_amp or args.version_prefix):
+            amp.build_amp(lang, args, cfg)
+
         if not args.skip_single_page:
-            build_single_page_version(lang, args, raw_config.get('nav'), cfg)
+            single_page.build_single_page_version(lang, args, raw_config.get('nav'), cfg)
 
         mdx_clickhouse.PatchedMacrosPlugin.disabled = False
 
@@ -185,175 +189,13 @@ def build_for_lang(lang, args):
         raise SystemExit('\n' + str(e))
 
 
-def build_single_page_version(lang, args, nav, cfg):
-    logging.info(f'Building single page version for {lang}')
-    os.environ['SINGLE_PAGE'] = '1'
-    extra = cfg.data['extra']
-    extra['single_page'] = True
-
-    with util.autoremoved_file(os.path.join(args.docs_dir, lang, 'single.md')) as single_md:
-        concatenate(lang, args.docs_dir, single_md, nav)
-
-        with util.temp_dir() as site_temp:
-            with util.temp_dir() as docs_temp:
-                docs_src_lang = os.path.join(args.docs_dir, lang)
-                docs_temp_lang = os.path.join(docs_temp, lang)
-                shutil.copytree(docs_src_lang, docs_temp_lang)
-                for root, _, filenames in os.walk(docs_temp_lang):
-                    for filename in filenames:
-                        if filename != 'single.md' and filename.endswith('.md'):
-                            os.unlink(os.path.join(root, filename))
-
-                cfg.load_dict({
-                    'docs_dir': docs_temp_lang,
-                    'site_dir': site_temp,
-                    'extra': extra,
-                    'nav': [
-                        {cfg.data.get('site_name'): 'single.md'}
-                    ]
-                })
-
-                if not args.test_only:
-                    mkdocs_build.build(cfg)
-
-                    if args.version_prefix:
-                        single_page_output_path = os.path.join(args.docs_dir, args.docs_output_dir, args.version_prefix, lang, 'single')
-                    else:
-                        single_page_output_path = os.path.join(args.docs_dir, args.docs_output_dir, lang, 'single')
-
-                    if os.path.exists(single_page_output_path):
-                        shutil.rmtree(single_page_output_path)
-
-                    shutil.copytree(
-                        os.path.join(site_temp, 'single'),
-                        single_page_output_path
-                    )
-
-                    single_page_index_html = os.path.join(single_page_output_path, 'index.html')
-                    single_page_content_js = os.path.join(single_page_output_path, 'content.js')
-                    with open(single_page_index_html, 'r') as f:
-                        sp_prefix, sp_js, sp_suffix = f.read().split('<!-- BREAK -->')
-                    with open(single_page_index_html, 'w') as f:
-                        f.write(sp_prefix)
-                        f.write(sp_suffix)
-                    with open(single_page_content_js, 'w') as f:
-                        if args.minify:
-                            import jsmin
-                            sp_js = jsmin.jsmin(sp_js)
-                        f.write(sp_js)
-
-                logging.info(f'Re-building single page for {lang} pdf/test')
-                with util.temp_dir() as test_dir:
-                    extra['single_page'] = False
-                    cfg.load_dict({
-                        'docs_dir': docs_temp_lang,
-                        'site_dir': test_dir,
-                        'extra': extra,
-                        'nav': [
-                            {cfg.data.get('site_name'): 'single.md'}
-                        ]
-                    })
-                    mkdocs_build.build(cfg)
-
-                    css_in = ' '.join(website.get_css_in(args))
-                    js_in = ' '.join(website.get_js_in(args))
-                    subprocess.check_call(f'cat {css_in} > {test_dir}/css/base.css', shell=True)
-                    subprocess.check_call(f'cat {js_in} > {test_dir}/js/base.js', shell=True)
-                    if args.save_raw_single_page:
-                        shutil.copytree(test_dir, args.save_raw_single_page)
-
-                    if not args.version_prefix:  # maybe enable in future
-                        logging.info(f'Running tests for {lang}')
-                        test.test_single_page(
-                            os.path.join(test_dir, 'single', 'index.html'), lang)
-
-                    if not args.skip_pdf:
-                        single_page_index_html = os.path.join(test_dir, 'single', 'index.html')
-                        single_page_pdf = os.path.abspath(
-                            os.path.join(single_page_output_path, f'clickhouse_{lang}.pdf')
-                        )
-
-                        with open(single_page_index_html, 'r') as f:
-                            soup = bs4.BeautifulSoup(
-                                f.read(),
-                                features='html.parser'
-                            )
-                        soup_prefix = f'file://{test_dir}'
-                        for img in soup.findAll('img'):
-                            if img['src'].startswith('/'):
-                                img['src'] = soup_prefix + img['src']
-                        for script in soup.findAll('script'):
-                            script['src'] = soup_prefix + script['src'].split('?', 1)[0]
-                        for link in soup.findAll('link'):
-                            link['href'] = soup_prefix + link['href'].split('?', 1)[0]
-
-                        with open(single_page_index_html, 'w') as f:
-                            f.write(str(soup))
-
-                        create_pdf_command = [
-                            'wkhtmltopdf',
-                            '--print-media-type',
-                            '--log-level', 'warn',
-                            single_page_index_html, single_page_pdf
-                        ]
-
-                        logging.info(' '.join(create_pdf_command))
-                        subprocess.check_call(' '.join(create_pdf_command), shell=True)
-
-        logging.info(f'Finished building single page version for {lang}')
-
-
-def write_redirect_html(out_path, to_url):
-    out_dir = os.path.dirname(out_path)
-    try:
-        os.makedirs(out_dir)
-    except OSError:
-        pass
-    with open(out_path, 'w') as f:
-        f.write(f'''<!--[if IE 6]> Redirect: {to_url} <![endif]-->
-<!DOCTYPE HTML>
-<html lang="en-US">
-    <head>
-        <meta charset="UTF-8">
-        <meta http-equiv="refresh" content="0; url={to_url}">
-        <script type="text/javascript">
-            window.location.href = "{to_url}";
-        </script>
-        <title>Page Redirection</title>
-    </head>
-    <body>
-        If you are not redirected automatically, follow this <a href="{to_url}">link</a>.
-    </body>
-</html>''')
-
-
-def build_redirect_html(args, from_path, to_path):
-    for lang in args.lang.split(','):
-        out_path = os.path.join(
-            args.docs_output_dir, lang,
-            from_path.replace('/index.md', '/index.html').replace('.md', '/index.html')
-        )
-        version_prefix = f'/{args.version_prefix}/' if args.version_prefix else '/'
-        target_path = to_path.replace('/index.md', '/').replace('.md', '/')
-        to_url = f'/docs{version_prefix}{lang}/{target_path}'
-        to_url = to_url.strip()
-        write_redirect_html(out_path, to_url)
-
-
-def build_redirects(args):
-    with open(os.path.join(args.docs_dir, 'redirects.txt'), 'r') as f:
-        for line in f:
-            from_path, to_path = line.split(' ', 1)
-            build_redirect_html(args, from_path, to_path)
-
-
 def build_docs(args):
     tasks = []
     for lang in args.lang.split(','):
         if lang:
             tasks.append((lang, args,))
     util.run_function_in_parallel(build_for_lang, tasks, threads=False)
-    build_redirects(args)
+    redirects.build_redirects(args)
 
 
 def build(args):
@@ -373,19 +215,7 @@ def build(args):
     if not args.skip_website:
         website.process_benchmark_results(args)
         website.minify_website(args)
-
-    for static_redirect in [
-        ('benchmark.html', '/benchmark/dbms/'),
-        ('benchmark_hardware.html', '/benchmark/hardware/'),
-        ('tutorial.html', '/docs/en/getting_started/tutorial/',),
-        ('reference_en.html', '/docs/en/single/', ),
-        ('reference_ru.html', '/docs/ru/single/',),
-        ('docs/index.html', '/docs/en/',),
-    ]:
-        write_redirect_html(
-            os.path.join(args.output_dir, static_redirect[0]),
-            static_redirect[1]
-        )
+        redirects.build_static_redirects(args)
 
 
 if __name__ == '__main__':
@@ -400,10 +230,12 @@ if __name__ == '__main__':
     arg_parser.add_argument('--enable-stable-releases', action='store_true')
     arg_parser.add_argument('--stable-releases-limit', type=int, default='4')
     arg_parser.add_argument('--lts-releases-limit', type=int, default='2')
+    arg_parser.add_argument('--nav-limit', type=int, default='0')
     arg_parser.add_argument('--version-prefix', type=str, default='')
     arg_parser.add_argument('--is-stable-release', action='store_true')
     arg_parser.add_argument('--skip-multi-page', action='store_true')
     arg_parser.add_argument('--skip-single-page', action='store_true')
+    arg_parser.add_argument('--skip-amp', action='store_true')
     arg_parser.add_argument('--skip-pdf', action='store_true')
     arg_parser.add_argument('--skip-website', action='store_true')
     arg_parser.add_argument('--test-only', action='store_true')
@@ -436,6 +268,7 @@ if __name__ == '__main__':
         args.skip_multi_page = True
         args.skip_website = True
         args.skip_pdf = True
+        args.skip_amp = True
 
     from build import build
     build(args)
