+++ website/main.html
@@ -2,6 +2,7 @@
 
 {% set language = config.theme.language %}
 {% set direction = config.theme.direction %}
+{% set is_amp = config.extra.is_amp %}
 {% set single_page = config.extra.single_page %}
 {% set no_footer = True %}
 {% set og_type = 'article' %}
@@ -22,21 +23,26 @@
 
 {% set data_version = config.extra.version_prefix or 'master' %}
 {% set data_single_page = 'true' if config.extra.single_page else 'false' %}
-{% set extra_html_attrs = 'data-version="' + data_version + '" data-single-page="' + data_single_page + '"' %}
-{% set extra_body_attrs = 'data-spy="scroll" data-target="#toc" data-offset="80"' %}
 
-{% extends "templates/base.html" %}
+{% if is_amp %}
+    Hi!
+{% else %}
+    {% set extra_html_attrs = 'data-version="' + data_version + '" data-single-page="' + data_single_page + '"' %}
+    {% set extra_body_attrs = 'data-spy="scroll" data-target="#toc" data-offset="80"' %}
 
-{% block content %}
-{% include "templates/docs/nav.html" %}
-<div class="container-fluid{% if direction == 'rtl' %} text-right{% endif %}">
-    <div class="row justify-content-end{% if not single_page %} justify-content-xl-center{% endif %}">
-        {% include "templates/docs/sidebar.html" %}
-        {% include "templates/docs/content.html" %}
-        {% if not config.extra.single_page %}
-            {% include "templates/docs/toc.html" %}
-        {% endif %}
+    {% extends "templates/base.html" %}
+
+    {% block content %}
+    {% include "templates/docs/nav.html" %}
+    <div class="container-fluid{% if direction == 'rtl' %} text-right{% endif %}">
+        <div class="row justify-content-end{% if not single_page %} justify-content-xl-center{% endif %}">
+            {% include "templates/docs/sidebar.html" %}
+            {% include "templates/docs/content.html" %}
+            {% if not config.extra.single_page %}
+                {% include "templates/docs/toc.html" %}
+            {% endif %}
+        </div>
     </div>
-</div>
 
-{% endblock %}
+    {% endblock %}
+{% endif %}
